Move needle client to client generator and needle base to root folder
