+++ generators/client/needle-api/needle-client-angular.js
