+++ generators/client/needle-api/needle-client-i18n.js
