+++ generators/client/needle-api/needle-client-webpack.js
