+++ generators/client/needle-api/needle-client.js
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const needleBase = require('./needle-base');
+const needleBase = require('../../needle-base');
 
 const constants = require('../../generator-constants');
