+++ generators/needle-api.js
@@ -16,11 +16,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const Base = require('./internal/needle-api/needle-base');
-const ClientAngular = require('./internal/needle-api/needle-client-angular');
-const ClientReact = require('./internal/needle-api/needle-client-react');
-const ClientWebpack = require('./internal/needle-api/needle-client-webpack');
-const ClientI18n = require('./internal/needle-api/needle-client-i18n');
+const Base = require('./needle-base');
+const ClientAngular = require('./client/needle-api/needle-client-angular');
+const ClientReact = require('./client/needle-api/needle-client-react');
+const ClientWebpack = require('./client/needle-api/needle-client-webpack');
+const ClientI18n = require('./client/needle-api/needle-client-i18n');
 const ServerMaven = require('./server/needle-api/needle-server-maven');
 const ServerGradle = require('./server/needle-api/needle-server-gradle');
 const ServerCache = require('./server/needle-api/needle-server-cache');
